functions/table_update/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery==1.24.0
google-cloud-storage==1.26.0
